requirements.txt
gspread>=6.1.2
google-auth>=2.29.0
requests>=2.32.3
pyyaml>=6.0.1
beautifulsoup4>=4.12.3
streamlit>=1.35.0
pandas>=2.0.0
